+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Package "reinstall" is now performed not only if `Version` field is different
+  but also when `Architecture`, `LinkSignature`, `Depends`, `Conflicts` or
+  `Provides` are different.
 ## [67.0] - 2020-06-23
 ### Added
 - Warning for packages not verified against repository index because missing hash